# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ระบบจัดการการลาของพนักงาน — Backend API สำหรับยื่นใบลา อนุมัติ ปฏิเสธ และดูยอดวันลาคงเหลือ",
        "title": "Leave Management System API",
        "contact": {
            "name": "API Support",
            "email": "support@company.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "ยืนยันตัวตนด้วยอีเมลและรหัสผ่าน จะได้รับ JWT token กลับมา",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "description": "ข้อมูลสำหรับเข้าสู่ระบบ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูลที่ส่งกลับ (ถ้ามี)"
                },
                "message": {
                    "description": "ข้อความอธิบาย",
                    "type": "string"
                },
                "success": {
                    "description": "สถานะความสำเร็จ",
                    "type": "boolean"
                }
            }
        },
        "dto.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT access token",
                    "type": "string"
                },
                "user": {
                    "description": "ข้อมูลผู้ใช้",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    ]
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "รายละเอียดข้อผิดพลาด (ถ้ามี)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "ข้อความอธิบายข้อผิดพลาด",
                    "type": "string"
                },
                "success": {
                    "description": "สถานะ (false เสมอ)",
                    "type": "boolean"
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "อีเมล",
                    "type": "string"
                },
                "password": {
                    "description": "รหัสผ่าน",
                    "type": "string"
                }
            }
        },
...
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "วันที่สร้างบัญชี",
                    "type": "string"
                },
                "email": {
                    "description": "อีเมล",
                    "type": "string"
                },
                "first_name": {
                    "description": "ชื่อจริง",
                    "type": "string"
                },
                "full_name": {
                    "description": "ชื่อเต็ม",
                    "type": "string"
                },
                "last_name": {
                    "description": "นามสกุล",
                    "type": "string"
                },
                "role": {
                    "description": "บทบาท",
                    "type": "string"
                },
                "user_id": {
                    "description": "รหัสผู้ใช้ (UUID)",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "กรุณาใส่ Bearer token เช่น \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}